fixes inline checkbox
