+++ lib/bureaucrat/widgets/checkbox_input.rb
@@ -1,13 +1,8 @@
-require 'bureaucrat/widgets/widget'
+require 'bureaucrat/widgets/base_checkbox_input'
 
 module Bureaucrat
   module Widgets
-    class CheckboxInput < Widget
-      def initialize(attrs=nil, check_test=nil)
-        super(attrs)
-        @check_test = check_test || lambda {|v| make_bool(v)}
-      end
-
+    class CheckboxInput < BaseCheckboxInput
       def render(name, value, attrs=nil)
         final_attrs = build_attrs(attrs, type: 'checkbox', name: name.to_s)
 
@@ -23,28 +18,6 @@ module Bureaucrat
         end
 
         mark_safe("<input#{flatatt(final_attrs)} />")
-      end
-
-      def value_from_formdata(data, name)
-        if data.include?(name)
-          value = data[name]
-
-          if value.is_a?(String)
-            case value.downcase
-            when 'true' then true
-            when 'false' then false
-            else value
-            end
-          else
-            value
-          end
-        else
-          false
-        end
-      end
-
-      def has_changed(initial, data)
-        make_bool(initial) != make_bool(data)
       end
     end
   end
